support wrapping command
